chore (refs T22614): move report components to client/js
